+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Fill asynchronously fake /dev/urandom for games that read a large number of bytes (#310)
 * Check for multiple opened pcm, and only keep the last opened
 * Support multiple ALSA connections
+* SDL relative mouse movement should ignore window clipping
 
 ## [1.3.5] - 2019-11-26
 ### Added